+++ fixture/application.py
@@ -1,5 +1,6 @@
 from selenium.webdriver.firefox.webdriver import WebDriver
 
+from fixture.contact import ContactHelper
 from fixture.group import GroupHelper
 from fixture.session import SessionHelper
 
@@ -10,44 +11,11 @@ class Application:
         self.wd.implicitly_wait(60)
         self.session = SessionHelper(self)
         self.group = GroupHelper(self)
+        self.contact = ContactHelper(self)
 
     def open_home_page(self):
         wd = self.wd
         wd.get("http://localhost:8080/addressbook/")
-
-    def back_to_contacts_list(self):
-        wd = self.wd
-        wd.find_element_by_link_text("home").click()
-
-    def fill_contact_form(self, contact):
-        # fill contacts form
-        wd = self.wd
-        self.open_add_contact_page()
-        wd.find_element_by_name("firstname").click()
-        wd.find_element_by_name("firstname").clear()
-        wd.find_element_by_name("firstname").send_keys(contact.first_name)
-        wd.find_element_by_name("lastname").click()
-        wd.find_element_by_name("lastname").clear()
-        wd.find_element_by_name("lastname").send_keys(contact.last_name)
-        wd.find_element_by_name("nickname").click()
-        wd.find_element_by_name("nickname").clear()
-        wd.find_element_by_name("nickname").send_keys(contact.nick)
-        wd.find_element_by_name("home").click()
-        wd.find_element_by_name("home").clear()
-        wd.find_element_by_name("home").send_keys(contact.home_phone)
-        wd.find_element_by_name("mobile").click()
-        wd.find_element_by_name("mobile").clear()
-        wd.find_element_by_name("mobile").send_keys(contact.mobile_phone)
-        wd.find_element_by_name("email").click()
-        wd.find_element_by_name("email").clear()
-        wd.find_element_by_name("email").send_keys(contact.email)
-        wd.find_element_by_xpath("//div[@id='content']/form/input[21]").click()
-        self.back_to_contacts_list()
-
-    def open_add_contact_page(self):
-        # open add creation new contact
-        wd = self.wd
-        wd.find_element_by_link_text("add new").click()
 
     def destroy(self):
         self.wd.quit()
